pyproject.toml
[project]
name = "asb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.2",
    "langchain>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.0.1",
    "langchain-ollama>=1.0.0",
    "ollama>=0.6.0",
    "openai>=2.7.1",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "sentence-transformers>=5.1.2",
    "typer>=0.20.0,<1",
    "click>=8.1.7,<9",
    "langchain-openai>=1.0.2",
    "networkx>=3.5",
    "apscheduler>=3.11.1",
    "sqlite-utils>=3.38",
    "notion-client>=2.7.0",
    "streamlit>=1.51.0",
    "requests>=2.32.5",
]

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["asb*"]
exclude = ["data*"]

[project.scripts]
asb = "asb.main:app"
